+++ shell/platform/android/BUILD.gn
@@ -213,6 +213,7 @@ android_java_sources = [
   "io/flutter/plugin/common/StandardMethodCodec.java",
   "io/flutter/plugin/common/StringCodec.java",
   "io/flutter/plugin/editing/FlutterTextUtils.java",
+  "io/flutter/plugin/editing/ImeSyncDeferringInsetsCallback.java",
   "io/flutter/plugin/editing/InputConnectionAdaptor.java",
   "io/flutter/plugin/editing/TextInputPlugin.java",
   "io/flutter/plugin/localization/LocalizationPlugin.java",
